Refactor schemas to own folder + eslint rule
