+++ .github/workflows/staging.yml
@@ -19,7 +19,6 @@ jobs:
       postgres:
         image: postgres
         env:
-          DATABASE_URL: ${{ secrets.DATABASE_URL }}
           POSTGRES_USER: ${{ secrets.DB_USERNAME }}
           POSTGRES_PASSWORD: ${{ secrets.DB_PASSWORD }}
           POSTGRES_DB: ${{ secrets.DB_DATABASE }}
@@ -28,7 +27,6 @@ jobs:
           --health-retries 5
         ports:
           - 5433:5432
-
     env:
       DATABASE_URL: ${{ secrets.DATABASE_URL }}
     steps:
@@ -41,12 +39,11 @@ jobs:
       - name: prettier
         run: npm run test:format
       - name: migrate db
-        run: npx prisma db push
+        run: npx prisma migrate deploy
       - name: test
         run: npm run test:ci
       - name: build
         run: npm run build
-
   deploy:
     runs-on: ubuntu-22.04
     environment: staging
