+++ package.json
@@ -59,7 +59,6 @@
     "lint-staged": "^14.0.1",
     "prettier": "^3.0.3",
     "prettier-plugin-prisma": "^5.0.0",
-    "prisma": "^5.3.1",
-    "whatwg-fetch": "^3.6.19"
+    "prisma": "^5.3.1"
   }
 }
