+++ src/components/Navbar/index.tsx
@@ -1,6 +1,6 @@
 import { AppBar, Typography } from '@mui/material';
 import Link from 'next/link';
-import { SignOutButton } from '../Buttons/Buttons';
+import { SignOutButton } from '@/components/Buttons/Buttons';
 import { getServerSession } from 'next-auth/next';
 import { authOptions } from '@/lib/auth';
